Working SQL Server Schema V1.2
This is a working Schema for SQL Server 2008 Express with Replication. It will work on newer platforms. This Schema was created using the Microsoft SQL Administrator. The Name of this Test SQL Server is DEEPEDDY\DEEPEDDYEXPRESS and you will notice that all .ASPX pages refer to it. When you deploy you will have to change the administration strings inside each of the .ASPX pages or use the same name scheme. Will move this to IP localhost architecture in near future.